* 2019.06.x:
  MCR-2110 fixed some codacy issues
  MCR-2110 fixed pmd issue in mycore-webtools
  MCR-2110 fixed pmd issues in mycore-classbrowser
  MCR-2110 set default value of pmd.failOnViolation to true
  MCR-2110 fixed appending of characters
  MCR-2110 make use of import statements to access JDOM2 classes
  MCR-2110 fixed checkstyle issues in mycore-base
  MCR-2110 fix checkstyle issues in mycore-ifs
  MCR-2110 fix checkstyle issues in mycore-restapi
  MCR-2110 fixed checkstyle issues in mycore-pi
  MCR-2110 fixed checkstyle issues in mycore-user2
  MCR-2110 set default of checkstyle.failOnViolation=true
  MCR-2110 added new lines at EOF
  MCR-2110 removed redundant modifiers and code
  MCR-2110 fix NeedBraces errors of checkstyle
  MCR-2110 fixed doctype of checkstyle.xml
  MCR-2110 reformatted code with eclipse 2019-09
  MCR-2110 organized imports, fixed line endings
  MCR-2110 Cleanup code
  MCR-2016 remove dependency xml-apis:xmls-apis
  MCR-2109 mods: use date of parent for indexing

+++ mycore-mets/src/main/java/org/mycore/mets/servlets/MCRMETSServlet.java
@@ -102,7 +102,7 @@ public class MCRMETSServlet extends MCRServlet {
 
         long lastModified = Files.getLastModifiedTime(rootPath).toMillis();
 
-        MCRFrontendUtil.writeCacheHeaders(response, (long) CACHE_TIME, lastModified, useExpire);
+        MCRFrontendUtil.writeCacheHeaders(response, CACHE_TIME, lastModified, useExpire);
         long start = System.currentTimeMillis();
         MCRContent metsContent = getMetsSource(job, useExistingMets(request), derivate);
         MCRLayoutService.instance().doLayout(request, response, metsContent);
